pyproject.toml
[tool.poetry]
name = "multilingual-repro"
version = "0.1.0"
description = ""
authors = ["Serkan Hosca <serkan@hosca.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
django-modeltranslation = "^0.19.9"
django-stubs = {extras = ["compatible-mypy"], version = "^5.1.0"}
django-model-utils = "^5.0.0"
ruff = "^0.6.8"
ruff-lsp = "^0.0.57"
ipython = "^8.27.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.django-stubs]
django_settings_module = "settings"

[tool.mypy]
# Mypy configuration:
# https://mypy.readthedocs.io/en/latest/config_file.html
plugins = [
    "mypy_django_plugin.main",
]
files = ["**/*.py"]
allow_redefinition = false
check_untyped_defs = true
disallow_incomplete_defs = true
disallow_untyped_defs = true
explicit_package_bases = true
ignore_errors = false
ignore_missing_imports = true
local_partial_types = true
no_implicit_optional = true
no_implicit_reexport = true
show_column_numbers = true
show_error_codes = true
strict_equality = true
strict_optional = true
warn_no_return = true
warn_redundant_casts = true
warn_return_any = true
warn_unused_configs = true
warn_unused_ignores = true

[tool.ruff]
line-length = 120

[tool.ruff.lint]
select = [
    "A",    # flake8-builtins
    "E",    # pycodestyle
    "W",    # pycodestyle
    "F",    # pyflakes
    "C4",   # flake8-comprehensions
    "C90",  # mccabe
    "I",    # isort
    "FLY",  # flynt
    "UP",   # pyupgrade
    "T10",  # flake8-debugger
    "B",    # flake8-bugbear
    "RET",  # flake8-return
    "RUF",  # ruff
    "EXE",  # flake8-executable
    "SIM",  # flake8-simplify
    "INT",  # flake8-gettext
    "PLC",  # pylint convention
    "PLE",  # pylint error
    "PLR",  # pylint refactor
    "PLW",  # pylint warning
]
ignore = [
    "E501",
    "RUF012", # Mutable class attributes should be annotated with `typing.ClassVar`
    "PLR0911", # Too many return statements
    "PLR0912", # Too many branches
    "PLR0913",  # Too many arguments in function definition
    "PLR0915", # Too many statements
    "PLR2004", # Magic value used in comparison, consider replacing with a constant variable
    "PLC0105", # `TypeVar` name does not reflect its covariance
]
fixable = ["ALL"]
unfixable = []

[tool.ruff.lint.isort]
force-single-line = true
force-sort-within-sections = true
no-lines-before = ["local-folder"]
